terraform/src/cloud_functions/generate_audio_upload_tmp/requirements.txt
google-cloud-storage==2.0.0
google-auth==2.16.2
requests==2.27.1